Refatoração para entrega
