--- src/br/com/dio/desafio/dominio/Curso.java
@@ -1,32 +1,0 @@
-package br.com.dio.desafio.dominio;
-
-public class Curso extends Conteudo{
-
-    private int cargaHoraria;
-
-    @Override
-    public double calcularXp() {
-        return XP_PADRAO * cargaHoraria;
-    }
-
-    public Curso() {
-    }
-
-
-    public int getCargaHoraria() {
-        return cargaHoraria;
-    }
-
-    public void setCargaHoraria(int cargaHoraria) {
-        this.cargaHoraria = cargaHoraria;
-    }
-
-    @Override
-    public String toString() {
-        return "Curso{" +
-                "titulo='" + getTitulo() + '\'' +
-                ", descricao='" + getDescricao() + '\'' +
-                ", cargaHoraria=" + cargaHoraria +
-                '}';
-    }
-}
