+++ src/main/java/br/com/dio/desafio/dominio/Conteudo.java
@@ -1,0 +1,18 @@
+package br.com.dio.desafio.dominio;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+public abstract class Conteudo {
+
+    protected static final double XP_PADRAO = 10d;
+
+    private String titulo;
+    private String descricao;
+
+    public abstract double calcularXp();
+}
