+++ src/main/java/br/com/dio/desafio/dominio/Curso.java
@@ -1,0 +1,22 @@
+package br.com.dio.desafio.dominio;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString(callSuper=true)
+public class Curso extends Conteudo{
+
+    private int cargaHoraria;
+
+    @Override
+    public double calcularXp() {
+        return XP_PADRAO * cargaHoraria;
+    }
+
+}
