+++ src/main/java/br/com/dio/desafio/dominio/Mentoria.java
@@ -1,0 +1,24 @@
+package br.com.dio.desafio.dominio;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+import java.time.LocalDate;
+
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString(callSuper=true)
+public class Mentoria extends Conteudo{
+
+    private LocalDate data;
+
+    @Override
+    public double calcularXp() {
+        return XP_PADRAO + 20d;
+    }
+
+}
